% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateOverDispersion.R
\name{estimateOverDispersion}
\alias{estimateOverDispersion}
\title{Estimate Overdispersion}
\usage{
estimateOverDispersion(obs, exp, length)
}
\arguments{
\item{obs}{Observed number of mutations per region}

\item{exp}{Expected number of mutations per region}

\item{length}{Total length of bases in region \emph{accumulated} across samples}
}
\description{
Estimate Overdispersion
}

