% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutLoads}
\alias{mutLoads}
\title{Expected and observed mutation load}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19256 rows and 6 columns.}
\usage{
mutLoads
}
\description{
Expected and observed mutation load
}
\keyword{datasets}

