% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdDetect.R
\name{ncdDetect}
\alias{ncdDetect}
\title{ncdDetect}
\usage{
ncdDetect(predictions, scores, observed_score, od, N = 11)
}
\arguments{
\item{od}{The standard deviation of a random effect in logit space. The random effect is not added to the first column. Can be estimated using \code{\link{estimateOverDispersion}}}
}
\description{
Find regions with an excessive mutation load
indicating positive selection
}

