% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddlepoint.R
\name{saddlepoint}
\alias{saddlepoint}
\title{Saddlepoint Approximation}
\usage{
saddlepoint(t, dat, lattice = 1L, log = T)
}
\arguments{
\item{t}{Point where tail probability is evaluated (note this function is not vectorized over t)}

\item{dat}{A data.table with columns 'x', 'y', and 'probabilities'. Column 'x' is a numeric indicator unique for each random variable to be convoluted together. Column 'y' is the outcome value, and colum 'probabilities' contains the probability corresponding to each value of the 'y' column.}

\item{lattice}{Lattice size in minimal lattice}

\item{log}{Return log probability}
}
\description{
Let \eqn{X_1,X_2,\ldots X_N} be independent Bernouilli trials
with \eqn{X_i \sim p_i}.
We evaluate the probability
\deqn{P(s_1X_1+s_2X_2\ldots s_NX_N > t)}.
}
\examples{
saddlepoint(51, dat = data.table(x = rep(1:10000, each = 2),
                                 y = rep(c(1,0),10000),
                                 probability = rep(c(0.01,0.99), 100)))
pbinom(299, 10000, 0.01, lower.tail = F, log.p = T)
}

